requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.0
numpy>=1.24.0